Main function
This code takes input from  the user by calling  the enter data function and it also displays the data by calling  the display data function. It also utilizes nested structure in the entire code to manage information  about  different  students their courses  and departments  in  which  they belong in a learning  institution.